Flags compilation

+++ src/parse/parse_map.c
@@ -6,7 +6,7 @@
 /*   By: rlins <rlins@student.42sp.org.br>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/03/08 17:46:34 by rlins             #+#    #+#             */
-/*   Updated: 2023/04/03 09:06:42 by rlins            ###   ########.fr       */
+/*   Updated: 2023/04/03 10:43:35 by rlins            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -85,8 +85,8 @@ static bool	fill_map_tab(t_map_det *map_det, char **map_tab)
  */
 static int	get_map_width(t_map_det *map_det)
 {
-	int	len;
-	int	i;
+	size_t	len;
+	int		i;
 
 	i = map_det->start_i_map;
 	len = ft_strlen(map_det->file[i]);
